create useAutocomplete hook to refactor component
